+++ src/TerminalLogin/FormPage.js
@@ -9,51 +9,9 @@ import CustomInput from '../ReUsableComponents/Custominput'
 import CustomSelect from '../ReUsableComponents/CustomSelect'
 import VirtualKeyboard from '../ReUsableComponents/VirtualKeyboard'
 import {Alert, AlertTitle } from '@mui/material' 
+import { labelStyle, buttonBoxStyle, buttonStyle, mainInputStyle, formBoxStyle, HeaderBox, FormBox } from './constants'
 
-// Style properties
-const labelStyle = {
-	fontWeight: 700,
-	fontSize: "1rem",
-	whiteSpace: "normal",
-	width: "%100",
-	minWidth: "125px",
-	marginBlockStart: 2,
-  };
-  
-  const HeaderBox = styled(Box)(({ theme }) => ({
-	border: "1px solid #9cdb9e",
-	boxShadow: "0",
-	padding: 0,
-	margin: 0,
-	width: "100vw",
-	[theme.breakpoints.up("sm")]: {
-	  width: "80vw",
-	},
-	[theme.breakpoints.up("md")]: {
-	  width: "70vw",
-	},
-	minWidth: 350,
-  }));
-  
-  const FormBox = styled(Box)(({ theme }) => ({
-	border: "1px solid #9cdb9e",
-	boxShadow: "0",
-	width: "99vw",
-	[theme.breakpoints.up("sm")]: {
-	  width: "80vw",
-	},
-	[theme.breakpoints.up("md")]: {
-	  width: "70vw",
-	},
-	marginInlineStart: { xs: 1, sm: 0 },
-	minWidth: 380,
-	display: "flex",
-	flexDirection: "column",
-	justifyContent: "space-around",
-	alignContent: "space-around",
-	alignItems: "center",
-  }));
-  
+
 // Initial values for the form fields
 const initialValues = {
 	terminal: "",
@@ -91,35 +49,8 @@ const validationSchema = yup.object().shape({
 	.string()
 	.required()
 })
- // Style properties
-const formBoxStyle={
-    marginInlineEnd:1,
-    marginInlineStart:1,
-	width:{xs:"99vw",sm:"70vw"},
-    minWidth:"312px",
-    maxWidth:"500px",
-    display:"flex",
-    justifyContent:"space-between",
-	overflow:"hidden"
-}
-const mainInputStyle={
-	width:{xs:"%100",sm:"%100",md:"%70"},
-	margin:1,
-	display:"flex",
-	flexGrow:1,
-	flexBasis:0
-}
-const buttonBoxStyle={
-	width:"%100",
-	margin:2,
-	marginBlock:0.5,
-	display:"flex",
-	alignItems:"center",
-}
-const buttonStyle={
-	width: "%100",
-	flex:1
-}
+
+
 const FormPage = () => {
 	// Set the background color of the body
 	document.body.style.backgroundColor = "#c6ffc8";
@@ -190,33 +121,40 @@ const FormPage = () => {
 		}
 	  };
 	  
-		// Fetch data from the server on component mount
-	  useEffect(() => {
-		axios
-		  .post("/login")
-		  .then(() => axios.get("/user"))
-		  .then((res) => {
+	// Fetch data from the server on component mount
+	useEffect(() => {
+		const fetchData = async () => {
+		try {
+			const loginResponse = await axios.post("/login");
+			await axios.get("/user").then((res) => {
 			setData(res.data.LoginPage);
-		  });
-	  }, []);
-	  
-	  let terminalOptions;
-	  let ShiftOptions;
-	  
-	  //Extract specified data from data
-	  if (data !== "empty") {
-		// Extract terminal options from the data
-		terminalOptions = data.LoginInfo.data.map((obj) => {
-		  const { termName } = obj;
-		  return termName;
-		});
-	  
-		// Extract shift options from the data
-		ShiftOptions = data.ShiftInfo.data.map((obj) => {
-		  const { shiftCode, rgbColor } = obj;
-		  return { shiftCode, rgbColor };
-		})
-	  }
+			});
+		} catch (error) {
+			console.error("Error fetching data:", error);
+		}
+		};
+	
+		fetchData();
+	}, []);
+  
+  let terminalOptions = [];
+  let ShiftOptions = [];
+  
+  // Extract specified data from data
+  if (data !== "empty") {
+	// Extract terminal options from the data
+	terminalOptions = data.LoginInfo.data.map((obj) => {
+	  const { termName } = obj;
+	  return termName;
+	});
+  
+	// Extract shift options from the data
+	ShiftOptions = data.ShiftInfo.data.map((obj) => {
+	  const { shiftCode, rgbColor } = obj;
+	  return { shiftCode, rgbColor };
+	});
+  }
+  
 	  
 
 	return data== "empty" ? <h1>Loading...</h1> : (
